giscanner: Introduce Page, convert mallardwriter
Convert mallardwriter over to a new Page subclass, as an intermediate step.
Eventually, we will remove this subclass, and repurpose the MallardFormatter
and DocBookFormatter to act on pages.

The concept of a page is different for a DocBook, so we'll need to add some
accessors to select certain kinds of subpages (methods, properties, signals).